adding shared module for shared classes
