+++ modeling-acceptance-tests/pom.xml
@@ -13,52 +13,9 @@
 
   <dependencies>
     <dependency>
-      <groupId>net.serenity-bdd</groupId>
-      <artifactId>serenity-jbehave</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>net.serenity-bdd</groupId>
-      <artifactId>serenity-core</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>net.serenity-bdd</groupId>
-      <artifactId>serenity-spring</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.core</groupId>
-      <artifactId>jackson-databind</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.assertj</groupId>
-      <artifactId>assertj-core</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.github.openfeign</groupId>
-      <artifactId>feign-core</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.github.openfeign</groupId>
-      <artifactId>feign-gson</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.github.openfeign</groupId>
-      <artifactId>feign-jackson</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.github.openfeign.form</groupId>
-      <artifactId>feign-form</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework</groupId>
-      <artifactId>spring-context</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.hateoas</groupId>
-      <artifactId>spring-hateoas</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.cloud</groupId>
-      <artifactId>spring-cloud-netflix-core</artifactId>
+      <groupId>org.activiti.cloud</groupId>
+      <artifactId>shared-acceptance-tests</artifactId>
+      <version>${project.version}</version>
     </dependency>
   </dependencies>
 </project>
