+++ modeling-acceptance-tests/src/main/java/org/activiti/cloud/qa/steps/ModelingGroupsSteps.java
@@ -22,16 +22,22 @@ import java.util.UUID;
 import javax.annotation.PostConstruct;
 
 import net.thucydides.core.annotations.Step;
+import org.activiti.cloud.qa.config.BaseTestsConfigurationProperties;
+import org.activiti.cloud.qa.config.ModelingTestsConfigurationProperties;
 import org.activiti.cloud.qa.model.modeling.Group;
+import org.activiti.cloud.qa.rest.ModelingFeignConfiguration;
 import org.activiti.cloud.qa.rest.feign.EnableFeignContext;
+import org.activiti.cloud.qa.rest.feign.FeignConfiguration;
 import org.activiti.cloud.qa.service.ModelingGroupsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.Resource;
+import org.springframework.test.context.ContextConfiguration;
 
 /**
  * Modeling groups steps
  */
 @EnableFeignContext
+@ContextConfiguration(classes = {ModelingTestsConfigurationProperties.class, ModelingFeignConfiguration.class})
 public class ModelingGroupsSteps extends ModelingContextSteps<Group> {
 
     @Autowired
