--- runtime-acceptance-tests/src/main/java/org/activiti/cloud/qa/rest/feign/HalDecoder.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2018 Alfresco, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.activiti.cloud.qa.rest.feign;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import feign.jackson.JacksonDecoder;
-import org.springframework.cloud.netflix.feign.support.ResponseEntityDecoder;
-import org.springframework.hateoas.hal.Jackson2HalModule;
-
-/**
- * HAL decoder
- */
-public class HalDecoder extends ResponseEntityDecoder {
-
-    public HalDecoder() {
-        super(new JacksonDecoder(
-                new ObjectMapper()
-                        .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
-                        .registerModule(new Jackson2HalModule())));
-    }
-}
