--- runtime-acceptance-tests/src/main/java/org/activiti/cloud/qa/rest/feign/OAuth2FeignRequestInterceptor.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2018 Alfresco, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.activiti.cloud.qa.rest.feign;
-
-import feign.RequestInterceptor;
-import feign.RequestTemplate;
-import net.serenitybdd.core.Serenity;
-
-/**
- * Feign RequestInterceptor to add Bearer token in all headers
- */
-public class OAuth2FeignRequestInterceptor implements RequestInterceptor {
-
-    public static final String BEARER = "Bearer";
-
-    public static final String AUTHORIZATION = "Authorization";
-
-    @Override
-    public void apply(RequestTemplate template) {
-        template.header(AUTHORIZATION,
-                        String.format("%s %s",
-                                      BEARER,
-                                      Serenity.sessionVariableCalled("authToken")));
-    }
-}
