+++ runtime-acceptance-tests/src/main/java/org/activiti/cloud/qa/steps/AuditSteps.java
@@ -20,11 +20,14 @@ import java.util.Collection;
 import java.util.Map;
 
 import net.thucydides.core.annotations.Step;
+import org.activiti.cloud.qa.config.RuntimeTestsConfigurationProperties;
 import org.activiti.cloud.qa.model.Event;
 import org.activiti.cloud.qa.model.EventType;
+import org.activiti.cloud.qa.rest.RuntimeFeignConfiguration;
 import org.activiti.cloud.qa.rest.feign.EnableFeignContext;
 import org.activiti.cloud.qa.service.AuditService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.test.context.ContextConfiguration;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -32,6 +35,7 @@ import static org.assertj.core.api.Assertions.*;
  * Audit steps
  */
 @EnableFeignContext
+@ContextConfiguration(classes = {RuntimeTestsConfigurationProperties.class, RuntimeFeignConfiguration.class})
 public class AuditSteps {
 
     @Autowired
