+++ shared-acceptance-tests/src/main/java/org/activiti/cloud/qa/model/AuthToken.java
