+++ shared-acceptance-tests/src/main/java/org/activiti/cloud/qa/rest/DirtyContextHandler.java
