+++ shared-acceptance-tests/src/main/java/org/activiti/cloud/qa/rest/feign/FeignRestDataClient.java
