+++ shared-acceptance-tests/src/main/java/org/activiti/cloud/qa/rest/feign/OAuth2FeignRequestInterceptor.java
