+++ shared-acceptance-tests/src/main/java/org/activiti/cloud/qa/rest/httpclient/Serializer.java
