+++ shared-acceptance-tests/src/main/java/org/activiti/cloud/qa/service/AuthenticationService.java
