+++ shared-acceptance-tests/src/main/java/org/activiti/cloud/qa/steps/AuthenticationSteps.java
