+++ shared-acceptance-tests/src/main/java/org/activiti/cloud/qa/story/AuthenticationActions.java
